cookie rewrite optimization and more secure

+++ core-server/src/features/auth/actions/command/logout-command.tsx
@@ -1,20 +1,20 @@
 "use server";
 
-import { getCurrentUserOrRedirect } from "../../current-user";
-import { JWTService } from "../../jwt-service";
+import { redirect } from "next/navigation";
+import { GetCurrentUser_Server } from "../../current-user";
+import { JWTService } from "../../jwt/jwt-service";
 import DeleteAccountById from "./delete-account-by-id-command";
+import { HOME_ROUTE } from "@/app/routes";
 
 export async function LogoutCommand(): Promise<void> {
-  const currentUser = await getCurrentUserOrRedirect();
-
-  if (currentUser.isGuest) {
-    console.log("REMOVE GUEST USER");
-  }
+  const currentUser = await GetCurrentUser_Server();
 
   await JWTService.clearAuthCookies();
 
   // If it was a guest account, remove the account
-  if (currentUser.isGuest == true) {
+  if (currentUser && currentUser.isGuest == true) {
     await DeleteAccountById(currentUser.accountId);
   }
+
+  redirect(HOME_ROUTE);
 }
